+++ src/bittytax/conv/parsers/__init__.py
@@ -59,6 +59,7 @@ from . import (
     nexo,
     okx,
     paxful,
+    paypal,
     poloniex,
     qtrade,
     qtwallet,